+++ src/c.zig
@@ -16,6 +16,7 @@ pub const c = @cImport({
         @cInclude("signal.h");
         @cInclude("errno.h");
         @cInclude("sys/wait.h");
+        @cInclude("time.h");
     } else if (builtin.os.tag == .linux) {
         @cInclude("GLES3/gl3.h");
         @cInclude("pty.h");
@@ -27,5 +28,6 @@ pub const c = @cImport({
         @cInclude("signal.h");
         @cInclude("errno.h");
         @cInclude("sys/wait.h");
+        @cInclude("time.h");
     }
 });